+++ ExpandWorldData/compatibility/General.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using HarmonyLib;
-using UnityEngine;
 namespace ExpandWorldData;
 
 #pragma warning disable IDE0051