+++ ExpandWorldData/data/WorldData.cs
@@ -1,6 +1,5 @@
 using System.ComponentModel;
 using Service;
-using YamlDotNet.Serialization;
 namespace ExpandWorldData;
 
 public class WorldData