+++ ExpandWorldData/features/BiomeHeat.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-using System.Reflection.Emit;
 using HarmonyLib;
 
 namespace ExpandWorldData;