+++ ExpandWorldData/features/Height.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using System.Reflection.Emit;
 using HarmonyLib;
-using Service;
 using UnityEngine;
 
 namespace ExpandWorldData;
@@ -52,9 +51,9 @@ public class GetAshlandsHeight
   private static double WidthRestriction = DefaultWidthRestriction;
   private static readonly double DefaultLengthRestriction = 1000f;
   private static double LengthRestriction = DefaultLengthRestriction;
-  public static bool Patch(Harmony harmony, double widthRestriction, double lengthRestriction)
+  public static void Patch(Harmony harmony, double widthRestriction, double lengthRestriction)
   {
-    if (WidthRestriction == widthRestriction && LengthRestriction == lengthRestriction) return false;
+    if (WidthRestriction == widthRestriction && LengthRestriction == lengthRestriction) return;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.GetAshlandsHeight));
     var transpiler = AccessTools.Method(typeof(GetAshlandsHeight), nameof(Transpiler));
     WidthRestriction = widthRestriction;
@@ -62,8 +61,6 @@ public class GetAshlandsHeight
     harmony.Unpatch(method, transpiler);
     if (WidthRestriction != DefaultWidthRestriction || LengthRestriction != DefaultLengthRestriction)
       harmony.Patch(method, transpiler: new HarmonyMethod(transpiler));
-
-    return true;
   }
 
   static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions)
@@ -80,23 +77,16 @@ public class GetAshlandsHeight
 public class CreateAshlandsGap
 {
   private static bool IsPatched = false;
-  public static bool Patch(Harmony harmony, bool doPatch)
+  public static void Patch(Harmony harmony, bool doPatch)
   {
-    if (IsPatched == doPatch) return false;
+    if (IsPatched == doPatch) return;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.CreateAshlandsGap));
     var prefix = AccessTools.Method(typeof(CreateAshlandsGap), nameof(DisableGap));
     IsPatched = doPatch;
     if (doPatch)
-    {
-      Log.Info("Patching CreateAshlandsGap");
       harmony.Patch(method, prefix: new HarmonyMethod(prefix));
-    }
     else
-    {
-      Log.Info("Unpatching CreateAshlandsGap");
       harmony.Unpatch(method, prefix);
-    }
-    return true;
   }
 
   static bool DisableGap(ref double __result)
@@ -109,23 +99,16 @@ public class CreateAshlandsGap
 public class CreateDeepNorthGap
 {
   private static bool IsPatched = false;
-  public static bool Patch(Harmony harmony, bool doPatch)
+  public static void Patch(Harmony harmony, bool doPatch)
   {
-    if (IsPatched == doPatch) return false;
+    if (IsPatched == doPatch) return;
     var method = AccessTools.Method(typeof(WorldGenerator), nameof(WorldGenerator.CreateDeepNorthGap));
     var prefix = AccessTools.Method(typeof(CreateAshlandsGap), nameof(DisableGap));
     IsPatched = doPatch;
     if (doPatch)
-    {
-      Log.Info("Patching CreateDeepNorthGap");
       harmony.Patch(method, prefix: new HarmonyMethod(prefix));
-    }
     else
-    {
-      Log.Info("Unpatching CreateDeepNorthGap");
       harmony.Unpatch(method, prefix);
-    }
-    return true;
   }
 
   static bool DisableGap(ref double __result)